requirements.txt
numpy==1.24.2
streamlit==1.20.0
pandas==1.5.3
plotly==5.13.1
plotly-express==0.4.1
Pillow==9.4.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
xlrd<2.0
openpyxl==3.1.2